+++ homepage/models.py
@@ -1,17 +1,15 @@
-import datetime
-
 from django.db import models
 from django.urls import reverse
+from django.utils.timezone import now
+from django.utils.translation import gettext_lazy as _
 
 from grid.models import Grid
 from package.models import BaseModel, Package
-from django.utils.translation import gettext_lazy as _
 
 
 class RotatorManager(models.Manager):
     def get_current(self):
-        now = datetime.datetime.now()
-        return self.get_queryset().filter(start_date__lte=now, end_date__gte=now)
+        return self.get_queryset().filter(start_date__lte=now(), end_date__gte=now())
 
 
 class Dpotw(BaseModel):