+++ package/repos/github.py
@@ -1,6 +1,7 @@
 from time import sleep
 
 from django.conf import settings
+from django.utils import timezone
 
 from github3 import GitHub, login
 
@@ -42,16 +43,23 @@ class GitHubHandler(BaseHandler):
         if repo is None:
             return package
 
-        # package.repo_watchers = repo.watchers_count
-        package.repo_watchers = repo.watchers
-        # package.repo_forks = repo.forks_count
-        package.repo_forks = repo.forks
+        # import pytest
+        # pytest.set_trace()
+
+        # package.repo_watchers = repo.watchers
+        package.repo_watchers = repo.watchers_count
+
+        if repo.archived:
+            package.date_repo_archived = timezone.now()
+
+        # package.repo_forks = repo.forks
+        package.repo_forks = repo.forks_count
+
         package.repo_description = repo.description
         # repo.stargazers_count
 
         contributors = []
-        # for contributor in repo.contributors():
-        for contributor in repo.iter_contributors():
+        for contributor in repo.contributors():
             contributors.append(contributor.login)
             self.manage_ratelimit()
 
@@ -69,8 +77,7 @@ class GitHubHandler(BaseHandler):
 
         from package.models import Commit  # Added here to avoid circular imports
 
-        # for commit in repo.commits():
-        for commit in repo.iter_commits():
+        for commit in repo.commits():
             self.manage_ratelimit()
             try:
                 commit_record, created = Commit.objects.get_or_create(